move upload methods back to media manager
